# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "961 Apart Docs",
        "description": "Description to Api",
        "version": "2.0"
    },
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Register",
                "operationId": "00bb7b7f2b364cea87cf66200f7ab7fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "password",
                                    "password_confirmation",
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "User phone number",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm Password",
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Last name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": {
                                        "user": {
                                            "first_name": "Vasya",
                                            "last_name": "Pupkin",
                                            "phone": "77076753444",
                                            "id": 3
                                        },
                                        "token": "2|bWbibKB1dgaTF5BCHi8YAPfiTqpU8bQyC34GB9D8d4f2221a"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "090be67f68e00aa334fe66e4e6b896d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "User phone number",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": {
                                        "user": {
                                            "first_name": "Vasya",
                                            "last_name": "Pupkin",
                                            "phone": "77076753444",
                                            "id": 3
                                        },
                                        "token": "2|bWbibKB1dgaTF5BCHi8YAPfiTqpU8bQyC34GB9D8d4f2221a"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/admin/address/all": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get all",
                "operationId": "109dd7e175900463fb4efadcd8a20ca6",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": [
                                        {
                                            "address": "",
                                            "longitude": "42.3",
                                            "latitude": "69.123"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/admin/address/{buildingId}": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Find by Building Id",
                "operationId": "ef71c9c67ea5fc12e27198543e895c0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Building id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Success",
                                    "data": {
                                        "address": "",
                                        "longitude": "42.3",
                                        "latitude": "69.123"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/address/{buildingId}/update": {
            "put": {
                "tags": [
                    "Address"
                ],
                "summary": "Update Address By Building",
                "operationId": "597f1aa794181da6e75fcce81734305c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Building id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successfully"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/address/{buildingId}/delete": {
            "delete": {
                "tags": [
                    "Address"
                ],
                "summary": "Delete Address by BuildingId",
                "operationId": "ece0bf0c8012e334b58e9fab3c97c0d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "BuildingId",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
...
}